requirements.txt
torch
flask
openai-whisper
moviepy
fpdf
gunicorn
imageio_ffmpeg
